Break apart `remote[_test].go`.
These files were starting to feel large in my previous change, but I didn't want to combine a bunch of functionality with breaking everything apart, so I left it large.  This change is just moving logic into separate files.

Related: https://github.com/sigstore/cosign/issues/666
